Add Alertmanager
This adds code to also set up alertmanager. In order for this to work,
it needs a Slack API URL, which is a secret. So, this encrypts that
value using ansible-vault. The secret it stored in a vault_password file
on the UC ansible host. Each site will have to encrypt its own secrets
with its own password.