Add mysql exporter
This adds a MySQL exporter. There is some complexity around networking
because MySQL is assumed to be running in the "host", which means the
container needs to know that IP address to break out of the Docker
network namespace. This would be much easier if Docker-for-linux
supported "host.docker.local" resolution, but [they don't][1].

[1]: https://github.com/docker/for-linux/issues/264