Allow whitelist of users to push to registry
Ideally we allow everybody in the Chameleon group to do this. That's a
bit annoying to do with the current auth set up on the Apache proxy.
It's definitely do-able, but involves dropping a(nother) service user
that can read project associations, which can then be used by an
additional external auth script that looks at groups for a user[1].

[1]: https://github.com/haegar/mod-auth-external/wiki/ConfigApache24#22-external-group-checkers