Simplify prometheus config generation
Currently we have a lot of boilerplate when setting up prometheus static
configs. This simplifies that a bunch by pulling the internal addresses
out to a fact first, then creating a 'jobs' structure that associates a
scrape job with an inventory group. The template references ansible
variables directly instead of relying on some magic vars being placed on
the template context.