Move snmp configuration out of neutron var
The SNMP configuration for the exporter caused the juniper switch to
start showing up in the neutron genericswitch configuration. But, there
is no support for managing the juniper switch currently, nor do we want
to (or even have that level of access.)

Moving the switch informatino to the more generic switch_configs solves
this. As already noted in the config, we need to somehow express the
switch configuration in a way where Neutron's view of it can be derived,
to reduce boilerplate.