[chi_tacc] Fix up monitoring
This fixes up the TACC inventory to work with the Prometheus monitoring
playbook again. The main changes are ensuring the network_interfaces are
set on the hosts, and also telling Prometheus to spin up on the control
node rather than the monitoring node.

This also brings the other TACC nodes in to the fold. It is a bit weird
to have a dedicated database node, but we do.