[prometheus] Make instance launch alert less noisy
Look at all projects and sum and see if the rate in the past hour is
above 3. This looks like it catches legit fail cases without triggering
for a bad launch attempt. We may have to raise the threshold. It would
be nice to have a # of successful launches (or totals) to compare
against.