Add support for post-deploy step (#29)
There is a generic post-deploy step which runs for any site, when you
run `./cc-ansible post-deploy` (this is also a KA step). It will first
invoke what KA would normally do, then call an extra play of the user's
choice (we set this to be playbooks/post_deploy.yml). That playbook
additionally supports chain-loading a site-specific playbook.

The generic playbook will set up the Ironic provisioning network, the
baremetal flavor, and other things needed for Ironic to work. It also
sets up NAT rules for Corsa switches (if any). It also ensures that the
latest ramdisk/kernel are available as images.