[neutron] Update policy for Rocky
This updates our policy override to include the default policies for
Neutron concatenated with the policies for the FWaaS extension. It is
currently necessary to concatenate them because the policy-as-code was
only merged in Neutron in Stein. When we are on Stein we can probably
remove these overrides, as the defaults are fine.

This policy has no deviations from the defaults.