[prometheus] Fix broken default config
When the additional service args were added, they had the unintended
effect of overriding any existing defaults that Prometheus had in its
CMD section of the Dockerfile. This meant that, amongst other things, it
couldn't find the prometheus configuration.

This is sort-of fixed in later versions of Prometheus past 2.6, where it
symlinks most things in to /etc/prometheus so that default settings work
even if they are not explicitly defined as flags.