Automatically update dependencies (#31)
This removes the Makefile as a way of keeping things updated, instead
using the invocation of cc-ansible itself as a sort of 'self-update'
mechanism. It will attempt to detect and fix the following:

- Missing virtualenv
- Outdated pip packages
- Outdated Ansible Galaxy roles
- Outdated Kolla-Ansible fork

We vendor Kolla-Ansible now so that we can properly detect when it gets
out of sync (the submodule should detect drift between the origin and
the checkout.)

This also removes the Makefile and some other dead code that we don't
need anymore, now that this exists.