Cleanup cron notifications on backup scripts (#45)
Instead of defining the same notification script in multiple places, all
of which will override others when they run on the same host, pull that
part out to a common task and include it from the other roles.

Also updates the notification script to include some helpful information
e.g. the type and help text for the metric, which Prometheus needs in
order to make sense of this.