Fix issue with init config generation
Fixes an issue where `cc-ansible init` will fail due to the site-config
not being generated properly from the example. The `mkdir -p` call was
generating the target directory before the copy occurred, so `cp`
interpreted that it should copy the source (site-config.example) within
the target, as opposed to creating the directory. End result was that
the site-config folder ended up having the site-config.example directory
nested within. Fix up so that only the _parent_ directory for the
site-config directory is created lazily.