[prometheus] Support node_exporter only hosts
On hosts that should only have the node exporter installed (it's a
static binary, so it can really be installed everywhere), it should not
be required to have Docker at all. This loosens up that requirement. We
see this mostly with Ceph, which comes with its own exporter and is
often on nodes w/ nothing else installed.