Update cc-ansible to pull KA from checkout tag (#52)
Instead of using a Git submodule, the script will now attempt to install
the package via pip using a GitHub link. To avoid doing extra work, we
try to check the refs and only re-install if the refs no longer match.

This allows you to specify different versions of Kolla-Ansible at
runtime, which allows this repo to serve multiple openstack releases
simultaneously (for example.)

This is aided by support for a new file .env in a site config, which can
specify arbitrary env variables.