Install virtualenv with system-site-packages
Virtualenv is currently not installed with system-site-packages to avoid
conflicts with libraries such as the cryptography library, which has
been seen on CentOS7. However, if pip is not installed with this method,
it has no access to 'yum', and so cannot install packages. Not good.

This approach installs the virtualenv using site-packages inherited from
the system, but then also forces upgrades/re-installs for any new
packages defined in requirements.txt or elsewhere.