Ensure virtualenv copy of ansible runs
If ansible is installed on the host machine, it can take priority in
some cases. This messes up lots of stuff, notably custom filters in
kolla_ansible (because it cannot find the module, which is installed in
a virtualenv.) The main fix is to ensure we invoke it with an absolute
path, but adding --force-reinstall also ensures that we have a more or
less complete copy of everything we need inside the virtualenv.