Add Docker service action plugin
We install a lot of non-KA docker services: Docker Registry, JupyterHub,
Redfish agent, and soon Keycloak. Much of this work is boilerplate. This
adds a new action_plugin in Ansible that can handle most of the work of
setting up a systemd service that runs a Docker container. It just does
a 'template' + 'service' invocation and supports some common Docker
args. Turns out action_plugins are a bit complicated to write, though.