[prometheus] Rename role
This role name clashes with the KA role of the same name. Long term we
should fold all of this functionality directly in to KA; short term, we
namespace it with a chameleon_ prefix so if we invoke the role, we don't
run the risk of accidentally invoking the KA one.