Don't move KA roles when invoking cc-ansible
In the past, to avoid collisions with KA roles, the entire roles
directory was renamed during play execution, so Ansible would not be
able to find them easily. This however prevented invoking KA roles if
you wanted to. We have now prefixed all custom roles that have the same
name as a KA role, so we can remove that behavior now.