[allocations] Portal-compatible deployment
This sets up the allocation service to live alongside the portal
infrastructure (portal Django app and resource API). To do so, it adds a
few additional configuration options which default to off, notably
enable_portal. When portal is enabled, the playbook will ensure an Nginx
configuration is included to route requests to the allocations service.

This also introduces monitoring of the Redis DB that supports the
allocations service.