Remove project_id from default openrc
Because the project_ids can be different across different sites, this
should reduce the probability that a user will get an error when they
try to use the same RC file across two different sites.

Before, all project_ids would be the same due to replication; when we
turn that off, that will not be guaranteed.