[docker_registry] Disable storage redirection
In later versions of the registry, the storage tiers can indicate that
clients can interface directly with them. It's unclear exactly when this
changed. But, it causes problems because there is no mechanism in the
Swift storage driver to differentiate b/w internal vs. external
requests. The registry was using the 'internal' Swift endpoint for
performance, but then was giving this endpoint to downstream clients,
which they could not use, as it's on a separate network.

To fix, disable redirection. Redirection is actually less performant
anyways, even if we used the public endpoint, because our registry is
running on its own host, and the Swift public endpoint is proxying
through the main control node. So using the public endpoint means we
send all traffic through the main node, instead of the storage node.