[jupyterhub] Convert to services + add Trovi (#88)
This adds Trovi configuration for JupyterHub, and also converts
JupyterHub to use the new services framework. When this is rolled out,
it will be served by HAProxy.

The plan would be to roll this out, and then do a DNS record switchover,
at which point Apache will stop serving the app, and HAProxy takes over.