Update IdP mapping for project nickname support
The Keystone mapping was only mapping project names, however, we also
need to ensure nicknames are synced. This is tricky for a lot of
reasons, but the core reason is that complex object claims cannot easily
be mapped by Keystone.

The other wrinkle is that, even with the primitive support that Keystone
_does_ have (now) for mapping complex claims, we actually _can't_ get
the data we need from the Apache mod_auth_openidc module in the event of
a JWT-verification auth. This type of auth is performed when you use
v3oidcpassword auth method, first requesting an access_token from the
IdP directly, and then exchanging that for a Keystone token. It also
affects v3oidcaccesstoken for the same reason: when authenticating via
access token, the Apache module doesn't have a mechanism for passing the
decoded JSON from the JWT as an env/header to the underlying
application. We are relying on this for websso logins to be able to
parse the complex "projects" claim that consists of a list of project
names and nicknames.

Adding a fallback mapping that uses a new "project_names" claim, which
is the simpler form of a project association returned by our Keycloak
IdP (and is only avaible on access_tokens), helps CLI users to at least
authenticate properly.

However, it does mean that, for nicknames to propagate, users need to
log in to Horizon at least once. This is probably fine, as they likely
need to log in in order to generate their openrc file. And, nicknames
are only really useful in the GUI anyways.