[horizon] Update top level user menu
This uses ``USER_MENU_LINKS``, which was introduced in the Queens release.
We used to link to the docs when clicking "Help"--now it will take the
user to the Help Desk on Portal. A separate "Documentation" button is now
here. This preserves the original openrc link as well.

This also removes the theme section from the menu; it is not great to
have users swap the theme as we have been using it to implement some things,
such as better error pages.