[horizon] Fix theme selection
Turns out, removing the default theme from Horizon's list breaks static
asset collection, which is required for our theme, which extends from
the default theme.

However, since Pike, there is a SELECTABLE_THEMES setting, which can
also do what we want, and in fact was created explicitly for this
use-case.