Force override distro python version
When K-A detects that the host is CentOS 8, it will assume the
containers are also running this, and override the distro python
version to 3.6. See ``distro_python_version_map``. This just
overrides that setting entirely, and is a blunt instrument for this
task. This should go when we have support for more distros, proper.