Make email relay use localhost
Having the email relay defined on the internal FQDN only makes sense
if HAProxy is somehow routing traffic to the SMTP relay locally, which
is not a behavior that is included by default in K-A or even our tooling.

Most SMTP servers bind to localhost by default, and b/c the containers run
in host networking, this should JustWork(tm). An external relay can be set
if need be.