Fixup issues with chameleon-specific roles
Many of the roles were missing "become: yes" for the sudo use-case.
The mariadb backup role was missing some defaults and also wasn't
creating the backup user.

For the heat download task, I made it so you can plop a google JSON
key (encrypted w/ vault) into the site config and it will use that.