Add openstackclient(s) to the venv installation
A common question is how to install/run the openstack client tools. Given that the admin-openrc is present in the site-config directory, users should default to using the venv we're creating.

Additional tools are present in climeleon, but as that targets desktop/local usage, it's not currently a good fit for running directly on the control node.