[Hack] set PASSWORDS_FILE to pass prechecks
kolla-ansible has added a precheck for items in the passwords file with
empty values. Likely due to either our invocation/modifications of
kolla-ansible, setting `--passwords` does not successfully set the value
of PASSWORDS_FILE.

In addition, this check is passing trivially! The task is scaning an
encrypted file, not the variables loaded from it. Therefore it is a noop
in the case where the passwords are encrypted by vault.