enable deployment of kvm compute node
This enables deployment of a nova-compute worker node to run KVM
instances. Add the desired host to the inventory/hosts file to deploy.

The post-deploy roles download some example images, and flavors. As
this feature is still in beta for Chameleon, they are set to admin-only
to avoid user confusion.