Wip/remove_ansible_features (#191)
* [Ansible] Remove Mitogen

Mitogen does not support ansible >= 2.10, which is required for
kolla-ansible. While the unreleased 0.3 version of Mitogen supports this in
theory, it has not had activity for over a year.

* [Ansible] disable fact caching

fact-caching only provides a minimal speedup, and is a common source of
user confusion. Users with large deployments can manually enable it if
desired.

* [Ansible] remove deprecated hash_behaviour option

hash_behavior was deprecated in ansible 2.9
it seems to have been re-added as of 5.0, but with many warnings about
"DO NOT USE". Things seem to work anyway.

Co-authored-by: Mark Powers <markpowers@uchicago.edu>