use upstream keystone IDP federation method
This adds static IDP metadata files for keystone, corresponding to
auth.chameleoncloud.org and auth.dev.chameleoncloud.org

By default, federation is enabled if keystone_idp_client_id is defined,
 and not null or false (truthy)

The `identity_provider_key` is set somewhat messily to correspond to the correct
hostname, so that the certificate can be loaded.

This needs a fair amount of cleanup, as upstream didn't use our (much cleaner)
way of fetching this data from the `.well_known` IDP endpoint.