Ensure post-deploy roles succeed
This fixes several issues with post-deploy. In particular, kolla-ansible
is very sensitive to the paths for roles and playbooks, to the extent
that we copy the playbooks and/or post-deploy yml files into the venv.

However, using "import_playbook" is currently proving intractible.

Therefore, this reverts to the method of having a single post-deploy
playbook calling roles, but also maintaining separate playbooks to
exercise the roles outside of post-deploy.