Refactor templating logic for clarity
This refactors the templating logic for doni workers and hardware types.
Appending to a list on condition, then joining the list, separates the
logic from syntax.

However, it may be preferrable move the logic to an ansible set_facts
role instead, as logic in a jinja template is an antipattern.