Install kubernetes dependency into ansible venv
This is hacky, but install the kubernetes dependency into the venv
created on the "controller node", rather than the chi-in-a-box venv.

Running the k8s role seems to pull from there, rather than from the
deploy host's venv.

The other option would be to delegate the k8s roles to the deploy host.