Fix fluentd logging when large backlog is present
If fluentd falls behind on sending logs to elasticsearch, eventually
the payload size will become too large for ES to accept.

This patch specifies `bulk_message_request_threshold`, causing fluentd to submit
multiple requests for payloads above 8mb.

This is done instead of increasing the max payload size, which can cause ES instability.

Additionally, the max header size for elasticsearch is increased from the default
of 16kb. Note that the path `elasticsearch/elasticsearch.yml` ensures that the change is applied
to the `elasticsearch` service, but NOT to `elasticsearch-curator`, as it has a different configuration format.