Added code to mount kubeconfig to hammer container (#276)
This change verifies that the kubeconfig_path ansible default is set by the user and fetches
the kubeconfig pointed to by that path. If the kubeconfig file is not present or not specified by the user; this code ensures that a placeholder kubeconfig is created for safe execution of the tasks.