use install_deps instead of auto-updating
auto-updating chi-in-a-box has caused frequent usability issues with
long installs, broken upstream python dependencies, and uncertainty on
which kolla-ansible version is in use.

This commit refactors all dependency installs into an "update_deps"
command, which must be invoked before other cc-ansible commands.
update_deps can be run repeatedly to update an existing venv.

Finally, we no longer edit the virtualenv by adding a symlink from
$VIRTUALENV/share to the kolla-ansible base directoy. Instead, in
the for editable installs of kolla-ansible (used by default in ciab), we
look up the actual install path by referencing the python dist-info.

This approach matches the one used in kolla-ansible, and is a
prerequisite to support installing kolla-ansible from a local checkout.