[neutron] don't enable l2population by default
the l2population mechanism is only necessary when using vxlan tenant
networks, which is not currently supported in chi-in-a-box. It is a no-op
for baremetal nodes with flat or vlan networks, and logs noisy warnings.

It was previously introduced here for chi@edge v1, which used vxlan networks
with Zun, but that was deprecated in v2.

NOTE: this will need to be revisited for KVM support, if vxlan networks are used